requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
accelerate==1.5.2
    # via stableguard (pyproject.toml)
aiofiles==23.2.1
    # via gradio
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.14
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   gradio
    #   httpx
    #   starlette
attrs==25.3.0
    # via aiohttp
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
datasets==3.5.0
    # via stableguard (pyproject.toml)
diffusers==0.32.2
    # via stableguard (pyproject.toml)
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
fastapi==0.115.12
    # via gradio
ffmpy==0.5.0
    # via gradio
filelock==3.18.0
    # via
    #   datasets
    #   diffusers
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.12.0
    # via
    #   datasets
    #   gradio-client
    #   huggingface-hub
    #   torch
gradio==5.23.1
    # via stableguard (pyproject.toml)
gradio-client==1.8.0
    # via gradio
groovy==0.1.2
    # via gradio
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   gradio
    #   gradio-client
    #   safehttpx
huggingface-hub==0.30.1
    # via
    #   accelerate
    #   datasets
    #   diffusers
    #   gradio
    #   gradio-client
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.6.1
    # via diffusers
invisible-watermark==0.2.0
    # via stableguard (pyproject.toml)
jinja2==3.1.6
    # via
    #   gradio
    #   torch
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   gradio
    #   jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.2.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
networkx==3.4.2
    # via torch
numpy==2.2.4
    # via
    #   accelerate
    #   datasets
    #   diffusers
    #   gradio
    #   invisible-watermark
    #   opencv-python
    #   pandas
    #   pywavelets
    #   torchvision
    #   transformers
opencv-python==4.11.0.86
    # via
    #   stableguard (pyproject.toml)
    #   invisible-watermark
orjson==3.10.16
    # via gradio
packaging==24.2
    # via
    #   accelerate
    #   datasets
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   transformers
pandas==2.2.3
    # via
    #   datasets
    #   gradio
piexif==1.1.3
    # via stableguard (pyproject.toml)
pillow==11.1.0
    # via
    #   stableguard (pyproject.toml)
    #   diffusers
    #   gradio
    #   invisible-watermark
    #   pillow-heif
    #   torchvision
pillow-avif-plugin==1.5.1
    # via stableguard (pyproject.toml)
pillow-heif==0.22.0
    # via stableguard (pyproject.toml)
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via accelerate
pyarrow==19.0.1
    # via datasets
pydantic==2.11.1
    # via
    #   fastapi
    #   gradio
pydantic-core==2.33.0
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-multipart==0.0.20
    # via gradio
pytz==2025.2
    # via pandas
pywavelets==1.8.0
    # via invisible-watermark
pyyaml==6.0.2
    # via
    #   accelerate
    #   datasets
    #   gradio
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via
    #   diffusers
    #   transformers
requests==2.32.4
    # via
    #   datasets
    #   diffusers
    #   huggingface-hub
    #   transformers
rich==14.0.0
    # via typer
ruff==0.11.2
    # via gradio
safehttpx==0.1.6
    # via gradio
safetensors==0.5.3
    # via
    #   accelerate
    #   diffusers
    #   transformers
semantic-version==2.10.0
    # via gradio
setuptools==78.1.0
    # via torch
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.1
    # via
    #   fastapi
    #   gradio
sympy==1.13.1
    # via torch
tokenizers==0.21.1
    # via transformers
tomlkit==0.13.2
    # via gradio
torch==2.5.1
    # via
    #   stableguard (pyproject.toml)
    #   accelerate
    #   invisible-watermark
    #   torchaudio
    #   torchvision
torchaudio==2.5.1
    # via stableguard (pyproject.toml)
torchvision==0.20.1
    # via stableguard (pyproject.toml)
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
transformers==4.50.3
    # via stableguard (pyproject.toml)
typer==0.15.2
    # via gradio
typing-extensions==4.13.0
    # via
    #   anyio
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via gradio
websockets==15.0.1
    # via gradio-client
xxhash==3.5.0
    # via datasets
yarl==1.18.3
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
